clean CommandQuery.AspNet.WebApi
