+++ src/CommandQuery.AspNet.WebApi/Internal/_InternalsVisibleTo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("CommandQuery.AspNet.WebApi.Tests")]
